requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.0
dask==2024.12.1
dask-expr==1.1.21
Faker==33.1.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
locket==1.0.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
partd==1.4.2
pip-autoremove==0.10.0
propcache==0.2.1
pyarrow==18.1.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
six==1.17.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
